pyproject.toml
[project]
name = "model-stacking-demo"
version = "0.1.0"
description = "Benchmarking setup for multi-model EfficientDet inference optimization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core ML
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "effdet>=0.4.1",
    
    # Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Load testing
    "locust>=2.20.0",
    
    # Image processing
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    
    # Utilities
    "httpx>=0.26.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
